+++ CHANGELOG.md
@@ -37,7 +37,7 @@
 - Add the support for performing swaps between cryptocurrencies through SwapKit
 - Fullscreen region select on mobile
 - Change portfolio chart default to year
-- iOS: price widget
+- iOS: live price widget with 24-hour chart
 
 ## v4.50.1
 - Fix a bug that would delay showing watch-only accounts.